requirements.txt
Pillow==9.4.0
PySide6==6.4.2
requests==2.28.1
retrying==1.3.4
rich==13.1.0
